// core/products/interfaces/product.interface.ts
import { User } from '@/core/auth/interface/user';

export interface Product {
    id: string;
    title: string;
    price: number;
    description: string;
    slug: string;
    stock: number;
    sizes: Size[];
    gender: Gender;
    tags: string[];
    images: string[];
    user?: User;
}

export enum Gender {
    Kid = 'kid',
    Men = 'men',
    Women = 'women',
}

export enum Size {
    L = 'L',
    M = 'M',
    S = 'S',
    Xl = 'XL',
    Xs = 'XS',
    Xxl = 'XXL',
    Xxxl = 'XXXL',
}